Add geopoint, geohash, and geohashRange types.
