+++ packages/geofire-common/src/index.ts
@@ -29,6 +29,10 @@ export const E2 = 0.00669447819799;
 // Cutoff for rounding errors on double calculations
 export const EPSILON = 1e-12;
 
+export type geopoint = [number, number];
+export type geohash = string;
+type geohashRange = [geohash, geohash];
+
 function log2(x: number): number {
   return Math.log(x) / Math.log(2);
 }
@@ -64,7 +68,7 @@ export function validateKey(key: string): void {
  *
  * @param location The [latitude, longitude] pair to be verified.
  */
-export function validateLocation(location: number[]): void {
+export function validateLocation(location: geopoint): void {
   let error: string;
 
   if (!Array.isArray(location)) {
@@ -96,7 +100,7 @@ export function validateLocation(location: number[]): void {
  *
  * @param geohash The geohash to be validated.
  */
-export function validateGeohash(geohash: string): void {
+export function validateGeohash(geohash: geohash): void {
   let error;
 
   if (typeof geohash !== 'string') {
@@ -139,7 +143,7 @@ export function degreesToRadians(degrees: number): number {
  * global default is used.
  * @returns The geohash of the inputted location.
  */
-export function geohashForLocation(location: number[], precision: number = GEOHASH_PRECISION): string {
+export function geohashForLocation(location: geopoint, precision: number = GEOHASH_PRECISION): geohash {
   validateLocation(location);
   if (typeof precision !== 'undefined') {
     if (typeof precision !== 'number' || isNaN(precision)) {
@@ -262,7 +266,7 @@ export function wrapLongitude(longitude: number): number {
  * @param size The size of the bounding box.
  * @returns The number of bits necessary for the geohash.
  */
-export function boundingBoxBits(coordinate: number[], size: number): number {
+export function boundingBoxBits(coordinate: geopoint, size: number): number {
   const latDeltaDegrees = size / METERS_PER_DEGREE_LATITUDE;
   const latitudeNorth = Math.min(90, coordinate[0] + latDeltaDegrees);
   const latitudeSouth = Math.max(-90, coordinate[0] - latDeltaDegrees);
@@ -278,10 +282,10 @@ export function boundingBoxBits(coordinate: number[], size: number): number {
  * to be prefixes of any geohash that lies within the circle.
  *
  * @param center The center given as [latitude, longitude].
- * @param radius The radius of the circle.
- * @returns The eight bounding box points.
+ * @param radius The radius of the circle in meters.
+ * @returns The center of the box, and the eight bounding box points.
  */
-export function boundingBoxCoordinates(center: number[], radius: number): number[][] {
+export function boundingBoxCoordinates(center: number[], radius: number): [geopoint, geopoint,geopoint,geopoint,geopoint,geopoint,geopoint,geopoint,geopoint] {
   const latDegrees = radius / METERS_PER_DEGREE_LATITUDE;
   const latitudeNorth = Math.min(90, center[0] + latDegrees);
   const latitudeSouth = Math.max(-90, center[0] - latDegrees);
@@ -308,7 +312,7 @@ export function boundingBoxCoordinates(center: number[], radius: number): number
  * @param bits The number of bits of precision.
  * @returns A [start, end] pair of geohashes.
  */
-export function geohashQuery(geohash: string, bits: number): string[] {
+export function geohashQuery(geohash: geohash, bits: number): geohashRange {
   validateGeohash(geohash);
   const precision = Math.ceil(bits / BITS_PER_CHAR);
   if (geohash.length < precision) {
@@ -337,13 +341,13 @@ export function geohashQuery(geohash: string, bits: number): string[] {
  * @param radius The radius of the circle.
  * @return An array of geohash query bounds, each containing a [start, end] pair.
  */
-export function geohashQueryBounds(center: number[], radius: number): string[][] {
+export function geohashQueryBounds(center: geopoint, radius: number): geohashRange[] {
   validateLocation(center);
   const queryBits = Math.max(1, boundingBoxBits(center, radius));
   const geohashPrecision = Math.ceil(queryBits / BITS_PER_CHAR);
   const coordinates = boundingBoxCoordinates(center, radius);
   const queries = coordinates.map((coordinate) => {
-    return geohashQuery(geohashForLocation(coordinate, geohashPrecision), queryBits);
+    return geohashQuery(geohashForLocation(coordinate as geopoint, geohashPrecision), queryBits);
   });
   // remove duplicates
   return queries.filter((query, index) => {
@@ -362,7 +366,7 @@ export function geohashQueryBounds(center: number[], radius: number): string[][]
  * @param location2 The [latitude, longitude] pair of the second location.
  * @returns The distance, in kilometers, between the inputted locations.
  */
-export function distanceBetween(location1: number[], location2: number[]): number {
+export function distanceBetween(location1: geopoint, location2: geopoint): number {
   validateLocation(location1);
   validateLocation(location2);
